Remove helpers and move to index.
